+++ dc.leaflet.js
@@ -1,5 +1,5 @@
 /*!
- *  dc.leaflet 0.5.1
+ *  dc.leaflet 0.5.2
  *  http://dc-js.github.io/dc.leaflet.js/
  *  Copyright 2014-2015 Boyan Yurukov and the dc.leaflet Developers
  *  https://github.com/dc-js/dc.leaflet.js/blob/master/AUTHORS
@@ -20,7 +20,7 @@
 'use strict';
 
 var dc_leaflet = {
-    version: '0.5.1'
+    version: '0.5.2'
 };
 
 dc_leaflet.leafletBase = function(Base) {
@@ -29,6 +29,7 @@ dc_leaflet.leafletBase = function(Base) {
     _chart.margins({left:0, top:0, right:0, bottom:0});
 
     var _map;
+    var _reuseMap = false;
 
     var _mapOptions=false;
     var _defaultCenter=false;
@@ -79,6 +80,8 @@ dc_leaflet.leafletBase = function(Base) {
             _chart.tiles()(_map);
             _chart._postRender();
         }
+        else if (_chart.map() && _reuseMap)
+            _chart._postRender();
         else
             console.warn("WARNING: Leaflet map already rendered.");
 
@@ -125,8 +128,13 @@ dc_leaflet.leafletBase = function(Base) {
         return _chart;
     };
 
-    _chart.map = function() {
-        return _map;
+    _chart.map = function(_) {
+        if (!arguments.length) {
+            return _map;
+        }
+        _map = _;
+        _reuseMap = true;
+        return _chart;
     };
 
     _chart._isMac = navigator.platform.toUpperCase().includes('MAC');
@@ -195,12 +203,12 @@ dc_leaflet.leafletBase = function(Base) {
 
     return _chart;
 };
-
 //Legend code adapted from http://leafletjs.com/examples/choropleth.html
 dc_leaflet.legend = function() {
     var _parent, _legend = {};
     var _leafletLegend = null;
     var _position = 'bottomleft';
+    var _legendTitle = false;
 
     _legend.parent = function (parent) {
         if(!arguments.length)
@@ -213,10 +221,17 @@ dc_leaflet.legend = function() {
         return L.Control.extend({
             options: {position: _position},
             onAdd: function (map) {
+                if (map.legend)
+                    map.legend.setContent("");
+                else
+                    map.legend = this;
                 this._div = L.DomUtil.create('div', 'info legend');
                 map.on('moveend',this._update,this);
                 this._update();
                 return this._div;
+            },
+            setContent: function (content) {
+                return this.getContainer().innerHTML = content;
             },
             _update: function () {
                 if (!_parent.colorDomain)
@@ -238,9 +253,12 @@ dc_leaflet.legend = function() {
                     }
 
                     var div = L.DomUtil.create('div', 'info legend');
+                    if (_legendTitle)
+                        this._div.innerHTML = "<strong>" + _legendTitle + "</strong><br/>";
+                    else
+                        this._div.innerHTML = ""; //reset so that legend is not plotted multiple times
                     // loop through our density intervals and generate a label with a colored
                     // square for each interval
-                    this._div.innerHTML = ""; //reset so that legend is not plotted multiple times
                     for (i = 0; i < grades.length; i++) {
                         this._div.innerHTML +=
                             '<i style="background:' + palette[i] + '"></i> ' +
@@ -255,6 +273,13 @@ dc_leaflet.legend = function() {
         if(!arguments.length)
             return _LegendClass;
         _LegendClass = LegendClass;
+        return _legend;
+    };
+
+    _legend.legendTitle = function(_) {
+        if(!arguments.length)
+            return _legendTitle;
+        _legendTitle = _;
         return _legend;
     };
 
@@ -288,7 +313,6 @@ dc_leaflet.legend = function() {
 
     return _legend;
 };
-
 dc_leaflet.markerChart = function(parent, chartGroup) {
     var _chart = dc_leaflet.leafletBase(dc.MarginMixin);
 
@@ -298,6 +322,8 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
     var _rebuildMarkers = false;
     var _brushOn = true;
     var _filterByArea = false;
+    var _clickEvent = false;
+    var _featureGroup = false;
 
     var _filter;
     var _innerFilter=false;
@@ -306,7 +332,10 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
     var _markerList = [];
     var _currentGroups=false;
 
-    _chart.renderTitle(true);
+    var _fitOnRender = true;
+    var _fitOnRedraw = false;
+    var _disableFitOnRedraw = false;
+    var _showMarkerTitle = true;
 
     var _location = function(d) {
         return _chart.keyAccessor()(d);
@@ -314,8 +343,8 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
 
     var _marker = function(d, map) {
         var marker = new L.Marker(_chart.toLocArray(_chart.locationAccessor()(d)), {
-            title: _chart.renderTitle() ? _chart.title()(d) : '',
-            alt: _chart.renderTitle() ? _chart.title()(d) : '',
+            title: _showMarkerTitle ? _chart.title()(d) : '',
+            alt: _showMarkerTitle ? _chart.title()(d) : '',
             icon: _icon(d, map),
             clickable: _chart.renderPopup() || (_chart.brushOn() && !_filterByArea),
             draggable: false
@@ -367,6 +396,7 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
         _layerGroup.clearLayers();
 
         var addList=[];
+        _featureGroup = false;
         groups.forEach(function(v, i) {
             var key = _chart.keyAccessor()(v);
             var marker = null;
@@ -387,6 +417,16 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
         if (_chart.cluster() && addList.length > 0) {
             _layerGroup.addLayers(addList);
         }
+
+        if (addList.length > 0) {
+            if (_fitOnRender || (_fitOnRedraw && !_disableFitOnRedraw)) {
+                _featureGroup = new L.featureGroup(addList);
+                _chart.map().fitBounds(_featureGroup.getBounds());
+            }
+        }
+
+        _disableFitOnRedraw = false;
+        _fitOnRender = false;
     };
 
     _chart.locationAccessor = function(_) {
@@ -405,6 +445,14 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
         return _chart;
     };
 
+    _chart.featureGroup = function(_) {
+        if (!arguments.length) {
+            return _featureGroup;
+        }
+        _featureGroup= _;
+        return _chart;
+    };
+
     _chart.icon = function(_) {
         if (!arguments.length) {
             return _icon;
@@ -418,6 +466,14 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
             return _popup;
         }
         _popup= _;
+        return _chart;
+    };
+
+    _chart.clickEvent = function(_) {
+        if (!arguments.length) {
+            return _clickEvent;
+        }
+        _clickEvent= _;
         return _chart;
     };
 
@@ -470,6 +526,30 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
         return _chart;
     };
 
+    _chart.fitOnRender = function(_) {
+        if (!arguments.length) {
+            return _fitOnRender;
+        }
+        _fitOnRender = _;
+        return _chart;
+    };
+
+    _chart.fitOnRedraw = function(_) {
+        if (!arguments.length) {
+            return _fitOnRedraw;
+        }
+        _fitOnRedraw = _;
+        return _chart;
+    };
+
+    _chart.showMarkerTitle = function(_) {
+        if (!arguments.length) {
+            return _showMarkerTitle;
+        }
+        _showMarkerTitle = _;
+        return _chart;
+    };
+
     _chart.markerGroup = function() {
         return _layerGroup;
     };
@@ -483,6 +563,9 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
         if (_chart.brushOn() && !_filterByArea) {
             marker.on("click", selectFilter);
         }
+        if (_clickEvent)
+            marker.on("click", _clickEvent)
+
         _markerList[k]=marker;
         return marker;
     };
@@ -555,7 +638,6 @@ dc_leaflet.markerChart = function(parent, chartGroup) {
 
     return _chart.anchor(parent, chartGroup);
 };
-
 dc_leaflet.choroplethChart = function(parent, chartGroup) {
     var _chart = dc_leaflet.leafletBase(dc.ColorMixin(dc.MarginMixin));
 
@@ -707,13 +789,24 @@ dc_leaflet.bubbleChart = function (parent, chartGroup) {
      * Private variables -- default values.
      * ####################################
      */
-    var _chart = dc_leaflet.leafletBase(dc.MarginMixin);
+    var _chart = dc_leaflet.leafletBase(dc.ColorMixin(dc.MarginMixin));
+    _chart.linearColors(['gray']);
     var _selectedColor = 'blue';
-    var _unselectedColor = 'gray';
+
+    var _unselectedColor = function(d) {
+        return _chart.getColor(d);
+    };
+    
+
+    var _renderPopup = true;
     var _layerGroup = false;
 
     var _location = function (d) {
         return _chart.keyAccessor()(d);
+    };
+
+    var _popup = function(d, feature) {
+        return _chart.title()(d);
     };
 
     var _r = d3.scaleLinear().domain([0, 100]);
@@ -728,13 +821,16 @@ dc_leaflet.bubbleChart = function (parent, chartGroup) {
 
         circle.setRadius(_chart.r()(_chart.valueAccessor()(d)));
         circle.on("mouseover", function (e) {
-            // TODO - Tooltips!
-            //console.log(_chart.title()(d));
+            if (_chart.renderPopup) this.openPopup();
         });
+        circle.on("mouseout", function (e) {
+            if (_chart.renderPopup) this.closePopup();
+        });
+
         var key = _chart.keyAccessor()(d);
         var isSelected = (-1 !== _chart.filters().indexOf(key));
 
-        circle.options.color = isSelected ? _chart.selectedColor() : _chart.unselectedColor();
+        circle.options.color = isSelected ? _chart.selectedColor() : _chart.unselectedColor()(d);
 
         return circle;
     };
@@ -801,9 +897,32 @@ dc_leaflet.bubbleChart = function (parent, chartGroup) {
         return _chart;
     };
 
+    _chart.popup = function(_) {
+        if (!arguments.length) {
+            return _popup;
+        }
+        _popup= _;
+        return _chart;
+    };
+
+    _chart.layerGroup = function() {
+        return _layerGroup;
+    };
+
+    _chart.renderPopup = function(_) {
+        if (!arguments.length) {
+            return _renderPopup;
+        }
+        _renderPopup = _;
+        return _chart;
+    };
+
     var createmarker = function (v, k) {
         var marker = _chart.marker()(v, _chart.map());
         marker.key = k;
+        if (_chart.renderPopup()) {
+            _chart.bindPopupWithMod(marker, _chart.popup()(v, marker));
+        }
         if (_chart.brushOn()) {
             marker.on("click", selectFilter);
         }
@@ -864,6 +983,10 @@ dc_leaflet.bubbleChart = function (parent, chartGroup) {
     }
 
     var selectFilter = function (e) {
+        if (!e.target) {
+            return;
+        }
+        var filter = e.target.key;
         L.DomEvent.stopPropagation(e);
         var filter = e.target.key;
         if (e.originalEvent.ctrlKey || e.originalEvent.metaKey) {
@@ -880,7 +1003,6 @@ dc_leaflet.bubbleChart = function (parent, chartGroup) {
 
     return _chart.anchor(parent, chartGroup);
 };
-
 dc_leaflet.d3 = d3;
 dc_leaflet.crossfilter = crossfilter;
 dc_leaflet.dc = dc;